requirements.txt
-e .
opencv-python
pillow
numpy
scikit-learn
requests
typer
rich
